+++ flexmeasures/api/v3_0/tests/test_sensor_schedules_fresh_db.py
@@ -74,6 +74,10 @@ def test_trigger_and_get_schedule(
         )
         print("Server responded with:\n%s" % trigger_schedule_response.json)
         assert trigger_schedule_response.status_code == 202
+        assert trigger_schedule_response.headers["Deprecation"] == "true"
+        assert 'rel="deprecation"' in trigger_schedule_response.headers["Link"]
+        assert "background-job-monitoring" in trigger_schedule_response.headers["Link"]
+        assert "deprecated-fields" not in trigger_schedule_response.json
         assert trigger_schedule_response.json["results-url"] == url_for(
             "SensorAPI:get_schedule",
             id=sensor.id,